requirements.txt
fastapi>=0.110
uvicorn>=0.24
pydantic>=2.6
pydantic-settings>=2.2
langgraph>=0.2
google-genai>=0.3
langgraph-checkpoint-sqlite>=0.2
python-multipart>=0.0.5
slowapi>=0.1.9
PyJWT>=2.8
pytest>=8.0
